Move pathfinder state to separate object
